+++ libs/async/test.c
@@ -37,13 +37,15 @@ static int am_async_reentrant(struct am_async *me, int *reent, int *state) {
     AM_ASYNC_BEGIN(me);
     if (*state == 0) {
         *state = 1;
-        AM_ASYNC_EXIT();
+        return AM_RC_DONE;
     }
     if (*state == 1) {
         *state = 2;
-        AM_ASYNC_EXIT();
+        return AM_RC_DONE;
     }
     AM_ASYNC_END();
+
+    return AM_RC_DONE;
 }
 
 static void test_async_local_continuation(void) {
@@ -70,6 +72,8 @@ static int am_async_empty(struct am_async *me, int *reent) {
     AM_ASYNC_BEGIN(me);
     ++(*reent);
     AM_ASYNC_END();
+
+    return AM_RC_DONE;
 }
 
 static void test_async_empty(void) {
@@ -89,6 +93,8 @@ static int am_async_wait_ready(struct am_async *me, int *reent, int ready) {
     AM_ASYNC_AWAIT(ready);
     ++(*reent);
     AM_ASYNC_END();
+
+    return AM_RC_DONE;
 }
 
 static void test_async_wait_ready(void) {
@@ -114,6 +120,8 @@ static int am_async_yield(struct am_async *me, int *state) {
     AM_ASYNC_YIELD();
     (*state) = 2;
     AM_ASYNC_END();
+
+    return AM_RC_DONE;
 }
 
 static void test_async_yield(void) {
@@ -140,11 +148,13 @@ static void test_async_yield(void) {
 static int am_async_exit(struct am_async *me, int *state) {
     AM_ASYNC_BEGIN(me);
     (*state) = 1;
-    AM_ASYNC_EXIT();
+    return AM_RC_DONE;
     AM_DISABLE_WARNING(AM_W_UNREACHABLE_CODE);
     (*state) = 2;
     AM_ENABLE_WARNING(AM_W_UNREACHABLE_CODE);
     AM_ASYNC_END();
+
+    return AM_RC_DONE;
 }
 
 static void test_async_exit(void) {
@@ -174,6 +184,8 @@ static enum am_rc am_async_call_1(struct am_async_chain *me) {
     AM_ASYNC_AWAIT(me->ready);
     me->foo = 1;
     AM_ASYNC_END();
+
+    return AM_RC_DONE;
 }
 
 static enum am_rc am_async_call_2(struct am_async_chain *me) {
@@ -181,6 +193,8 @@ static enum am_rc am_async_call_2(struct am_async_chain *me) {
     AM_ASYNC_AWAIT(me->ready);
     me->foo = 1;
     AM_ASYNC_END();
+
+    return AM_RC_DONE;
 }
 
 static void test_async_call_chain(void) {